Check if node has DOR service

+++ saas/dor/blueprint.py
@@ -100,6 +100,10 @@ class DORBlueprint:
 
     @request_manager.verify_request_body(search_body_specification)
     def search(self):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         body = request_manager.get_request_variable('body')
         patterns = body['patterns'] if 'patterns' in body else None
         owner_iid = body['owner_iid'] if 'owner_iid' in body else None
@@ -109,6 +113,10 @@ class DORBlueprint:
     @request_manager.verify_request_body(add_body_specification)
     @request_manager.verify_request_files(['attachment'])
     def add(self):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         body = request_manager.get_request_variable('body')
         files = request_manager.get_request_variable('files')
 
@@ -123,6 +131,10 @@ class DORBlueprint:
 
     @request_manager.verify_request_body(add_gpp_body_specification)
     def add_gpp(self):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         body = request_manager.get_request_variable('body')
 
         owner_iid = body['owner_iid']
@@ -135,10 +147,18 @@ class DORBlueprint:
 
     @request_manager.verify_authorisation_by_owner('obj_id')
     def delete(self, obj_id):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         status, result = self._node.dor.delete(obj_id)
         return jsonify(result), status
 
     def get_descriptor(self, obj_id):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         descriptor_path = self._node.dor.obj_descriptor_path(obj_id)
         if os.path.isfile(descriptor_path):
             descriptor = read_json_from_file(descriptor_path)
@@ -148,6 +168,10 @@ class DORBlueprint:
 
     @request_manager.verify_authorisation_by_owner('obj_id')
     def get_content(self, obj_id):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         # get the content hash for the data object
         record = self._node.db.get_object_by_id(obj_id)
         if not record:
@@ -158,6 +182,10 @@ class DORBlueprint:
         return send_from_directory(head, tail, as_attachment=True)
 
     def get_access_overview(self, obj_id):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         if not self._node.db.get_object_by_id(obj_id):
             return jsonify(f"{obj_id} not found"), 404
 
@@ -168,6 +196,10 @@ class DORBlueprint:
 
     @request_manager.verify_authorisation_by_owner('obj_id')
     def grant_access(self, obj_id, iid):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         if not self._node.db.get_object_by_id(obj_id):
             return jsonify(f"data object (id={obj_id}) not found"), 404
 
@@ -181,6 +213,10 @@ class DORBlueprint:
 
     @request_manager.verify_authorisation_by_owner('obj_id')
     def revoke_access(self, obj_id: str, iid: str):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         if not self._node.db.get_object_by_id(obj_id):
             return jsonify(f"data object (id={obj_id}) not found"), 404
 
@@ -193,6 +229,10 @@ class DORBlueprint:
         }), 200
 
     def get_owner(self, obj_id: str):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         owner = self._node.db.get_owner(obj_id)
         if owner:
             return jsonify({
@@ -205,6 +245,10 @@ class DORBlueprint:
     @request_manager.verify_request_body(transfer_ownership_body_specification)
     @request_manager.verify_authorisation_by_owner('obj_id')
     def transfer_ownership(self, obj_id: str):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         # get the record for this data object
         record = self._node.db.get_object_by_id(obj_id)
         if record is None:
@@ -224,11 +268,19 @@ class DORBlueprint:
         return jsonify({obj_id: owner.id}), 200
 
     def get_tags(self, obj_id: str):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         return jsonify(self._node.db.get_tags(obj_id)), 200
 
     @request_manager.verify_request_body(tags_body_specification)
     @request_manager.verify_authorisation_by_owner('obj_id')
     def update_tags(self, obj_id):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         body = request_manager.get_request_variable('body')
         self._node.db.update_tags(obj_id, body)
 
@@ -237,6 +289,10 @@ class DORBlueprint:
     @request_manager.verify_request_body(delete_tags_body_specification)
     @request_manager.verify_authorisation_by_owner('obj_id')
     def remove_tags(self, obj_id):
+        # does this node have a DOR?
+        if self._node.dor is None:
+            return jsonify("Node does not have a DOR service."), 404
+
         body = request_manager.get_request_variable('body')
         self._node.db.remove_tags(obj_id, body)
 
@@ -257,7 +313,7 @@ class DORProxy(EndpointProxy):
             body['owner_iid'] = owner_iid
 
         code, r = self.get('', body=body)
-        return r
+        return r if code == 200 else None
 
     def add_data_object(self, content_path, owner: Identity, access_restricted, content_encrypted,
                         data_type, data_format, created_by, created_t=None, recipe=None):
@@ -277,7 +333,7 @@ class DORProxy(EndpointProxy):
             body['descriptor']['recipe'] = recipe
 
         code, r = self.post('/add', body=body, attachment=content_path)
-        return (r['data_object_id'], r['descriptor']) if 'data_object_id' in r else None
+        return (r['data_object_id'], r['descriptor']) if code == 200 or code == 201 else None
 
     def add_gpp_data_object(self, source: str, commit_id: str, proc_path: str, proc_config: str, owner: Identity,
                             created_by, created_t=None, recipe=None, git_credentials: GithubCredentials = None):
@@ -309,7 +365,7 @@ class DORProxy(EndpointProxy):
 
         # execute post request and remove temp file afterwards
         code, r = self.post('/add-gpp', body=body)
-        return (r['data_object_id'], r['descriptor']) if 'data_object_id' in r else None
+        return (r['data_object_id'], r['descriptor']) if code == 200 or code == 201 else None
 
     def delete_data_object(self, obj_id, with_authorisation_by):
         code, r = self.delete(f"/{obj_id}", with_authorisation_by=with_authorisation_by)
@@ -325,19 +381,19 @@ class DORProxy(EndpointProxy):
 
     def get_access_overview(self, obj_id):
         code, r = self.get(f"/{obj_id}/access")
-        return r
+        return r if code == 200 else None
 
     def grant_access(self, obj_id: str, authority: Keystore, identity: Identity):
         code, r = self.post(f"/{obj_id}/access/{identity.id}", with_authorisation_by=authority)
-        return r
+        return r if code == 200 else None
 
     def revoke_access(self, obj_id: str, authority: Keystore, identity: Identity):
         code, r = self.delete(f"/{obj_id}/access/{identity.id}", with_authorisation_by=authority)
-        return r
+        return r if code == 200 else None
 
     def get_owner(self, obj_id: str):
         code, r = self.get(f"/{obj_id}/owner")
-        return r
+        return r if code == 200 else None
 
     def transfer_ownership(self, obj_id: str, authority: Keystore, new_owner: Identity, content_key: str = None):
         body = {
@@ -363,8 +419,8 @@ class DORProxy(EndpointProxy):
             })
 
         code, r = self.put(f"/{obj_id}/tags", body=body, with_authorisation_by=authority)
-        return r
+        return r if code == 200 else None
 
     def remove_tags(self, obj_id: str, authority: Keystore, keys: list):
         code, r = self.delete(f"/{obj_id}/tags", body=keys, with_authorisation_by=authority)
-        return r
+        return r if code == 200 else None
